change: 修改Host vs2015工程，使其集成entityx，eXtendedMirror

+++ OpenSource/eXtendedMirror-master/src/XM/Function.cpp
@@ -79,9 +79,9 @@ bool Function::before_(const Item& item) const
     if (!fullSignature_ || !other.fullSignature_)
         return false;
 
-    ushort paramN1 = params_.size();
-    ushort paramN2 = other.params_.size();
-    ushort paramN = std::min(paramN1, paramN2);
+    ushort paramN1 = (ushort)params_.size();
+    ushort paramN2 = (ushort)other.params_.size();
+    ushort paramN = (ushort)std::min(paramN1, paramN2);
     for (uint i = 0; i < paramN; i++)
     {
         if (params_[i]->type
